Support all OSM records
Records from OSM can (currently) be in the `address`, `venue` or
`street` layer (via the polylnes importer) layer.

For records from the OSM importer, [we changed the ID format](https://github.com/pelias/openstreetmap/pull/475)
so the `source_id` (the 3rd and final component of a Pelias GID) matches
exactly with the path on openstreetmap.org for a given record.

This change updates the matching code to support those cases.