Display full response body on error
The compare app has always had fairly cryptic messaging when there was
an error response from one of the Pelias servers.

It ignores any actual response content from Pelias, and instead displays
an error message ('failed to load json') as well as a bit of static,
synthesized JSON. Neither of these contain much information at all.

This PR changes around the behavior so that in the case of any
error-like response (really, a response that does not have the
`geocoding` section that Pelias generally returns), the actual response
body is displayed, which might help debugging.